refactor: introduce ChatWebSocketController
